test(scripts): extract lone-surrogate UTF-16 check and cover it
scripts/validate-raycast-feed.mjs guards the emitted Raycast JSON feed against
lone (unpaired) UTF-16 surrogates, but that check (stringHasLoneSurrogate) had no
unit tests. Move the pure function into scripts/lib/lone-surrogate.mjs - matching
the existing scripts/lib convention - and import it back.

- scripts/lib/lone-surrogate.mjs: stringHasLoneSurrogate.
- scripts/validate-raycast-feed.mjs: import the check; drop the local copy
  (assertNoLoneSurrogates still recurses through it).
- tests: cover BMP text / valid surrogate pair / empty string (false), a high
  surrogate not followed by a low one and a high surrogate at end-of-string, a
  standalone low surrogate, and a low surrogate that completes a pair. Branch
  coverage 100% (12/12).

No behavior change; the surrogate check is identical.
